Store Manager Service Towords Testing

+++ src/main/java/com/store_manager/services/OrderItemsService.java
@@ -22,7 +22,11 @@ public class OrderItemsService {
         IOrderItemsRepository orderItemsRepo,
         IOrderRepository orderRepo
     ){
-
+        if (orderRepo == null || orderItemsRepo == null || menuItemsRepo == null){
+            throw new IllegalArgumentException(
+                " orderRepo/orderItemRepo/menuItemsRepo parameter is null"
+            );
+        }
         this.menuItemsRepo   = menuItemsRepo;
         this.orderItemsRepo = orderItemsRepo;
         this.orderRepo = orderRepo;
@@ -32,40 +36,96 @@ public class OrderItemsService {
 
     /**
      * @brief Returns a list of all order items if order is valid otherwise null .
+     * @throws IllegalArgumentException If order is null
      */
     public List<OrderItemsModel> getAllOrderItemsFromOrder(OrderModel order){
+        if (order == null){
+            throw new IllegalArgumentException(
+                " order parameter is null"
+            );
+        }
         return orderItemsRepo.findByOrderId(order.getId());
     } 
 
     /**
      * @brief Returns from this order all items of the restaurant matched
-     * with the order items in the same order as the order items . If 
-     * no order item exists return's null .
+     * with the order items in the same order as the order items . 
+     * @param order The Order to a store
+     * @throws IllegalArgumentException If order is null
      */
     public List<MenuItemsModel> getAllMenuItemsOfOrder(OrderModel order){
+        if (order == null){
+            throw new IllegalArgumentException(
+                " order parameter is null"
+            );
+        }
+
         return orderItemsRepo.getMenuItemsByOrderItems(
             orderItemsRepo.findByOrderId(order.getId())
             );
     }
 
+    /**
+     * @brief Update a list of order items only if this operation is succesfull
+     * for all of them otherwise the operation is discarded .
+     * @returns A code that indicates if the operation is succesfull 
+     * 1. SUCCESS If everything is succesfull
+     * 2. FAILED_TO_WRITE If at least one items couldn't updated
+     * 3. NOT_FOUND If at least one items could not found on io 
+     * 4. IO_ERROR If io has more generic error
+     * @throws IllegalArgumentException If orderItems is null
+     */
     public ErrorCodes UpdateOrderItems(List<OrderItemsModel> orderItems){
+        if (orderItems == null){
+            throw new IllegalArgumentException(
+                " order items parameter is null"
+            );
+        }
+        ErrorCodes res= orderItemsRepo.UpdateOrderItems(orderItems);
+        
+        return res;
+    }
 
-        for( OrderItemsModel orderItem : orderItems){
-            orderItemsRepo.update(orderItem);
+    /**
+     * @brief Updates the order item from the order .
+     * @param orderItem The Order item to update .
+     * @return A code that represents if io operation happen succefully
+     * 1. SUCCESS If updated .
+     * 2. NOT_FOUND If orderItem could not be finded in io .
+     * 3. FAILED_TO_WRITE If io failed to write the order item .
+     * 4. IO_ERROR If io has a more generic error .
+     * @throws IllegalArgumentException If orderItem is null
+     */
+    public ErrorCodes UpdateOrderItem(OrderItemsModel orderItem){
+        if (orderItem == null){
+            throw new IllegalArgumentException(
+                " order item parameter is null"
+            );
         }
 
-        return ErrorCodes.SUCCESS;
-    }
-
-    public ErrorCodes UpdateOrderItem(OrderItemsModel orderItem){
         ErrorCodes res = orderItemsRepo.update(orderItem);
-        if (res != ErrorCodes.SUCCESS)
+        if (res != ErrorCodes.SUCCESS){
             return res;
+        }
+        
         return ErrorCodes.SUCCESS;
     }
 
+    /**
+     * @brief Deletes the order item from the order .
+     * @param orderItem The Order item to delete .
+     * @return A code that represents if io operation happen succefully
+     * 1. SUCCESS If deleted .
+     * 2. NOT_FOUND If orderItem could not be finded in io .
+     * 3. IO_ERROR If io has a more generic error .
+     * @throws IllegalArgumentException If orderItem is null
+     */
     public ErrorCodes DeleteOrderItem(OrderItemsModel orderItem){
-
+        if (orderItem == null){
+            throw new IllegalArgumentException(
+                " order item parameter is null"
+            );
+        }
         ErrorCodes res = orderItemsRepo.deleteById(orderItem.getId());
         if (res != ErrorCodes.SUCCESS){
             return res;
@@ -74,16 +134,36 @@ public class OrderItemsService {
         return ErrorCodes.SUCCESS;
     }
 
+
+    /**
+     * @brief Creates a new order based on menu item and quantity or Updates 
+     * the order that already exist's based on the quantity and menu item .
+     * @param menuItem The menu item connected to this order item
+     * @param OrderItemQuantity The additive quantity if equal order item already exist's
+     * or the total quantity if the new order item .
+     * @param order The Order that this order item will created on , basicly a way to 
+     * connect the store , customer to a order item is via the order .
+     * @note The order item consider equal if there is other order item with same menu item
+     * id and same order id .
+     * @throws  IllegalArgumentException If menuItem and/or order is null .
+     */
     public ErrorCodes CreateOrderItem(
         MenuItemsModel menuItem , 
         int orderItemQuantity , 
         OrderModel order){
         
+        if (order == null || menuItem == null){
+            throw new IllegalArgumentException(
+                " order and / or menuItem parameter is null"
+            );
+        }
+        
         OrderItemsModel newOrderItem = new OrderItemsModel();
         newOrderItem.setMenu_item_id(menuItem.getItem_id());
         newOrderItem.setQuantity(orderItemQuantity);
         newOrderItem.setOrder_id(order.getId());
-
+        
+        
         List<OrderItemsModel> inDBOrderItems = orderItemsRepo.findByOrderId(order.getId());
 
 
@@ -102,11 +182,12 @@ public class OrderItemsService {
             
             
         }
+
         if (!found){
             res = orderItemsRepo.save(newOrderItem);
         }
         else{
-            System.out.println("Common!!!");
+            
             res = orderItemsRepo.update(newOrderItem);
         }
